src/main/java/com/tta/geumgiri/member/service/MemberService.java
package com.tta.geumgiri.member.service;


import com.tta.geumgiri.member.domain.Member;
import com.tta.geumgiri.member.domain.dto.AddMemberRequest;
import com.tta.geumgiri.member.repository.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;

@RequiredArgsConstructor
@Service
public class MemberService {

    private final MemberRepository memberRepository;
    private final BCryptPasswordEncoder bCryptPasswordEncoder;

    public Long save(AddMemberRequest dto) throws EmailAlreadyExistsException {
        if (memberRepository.findByEmail(dto.getEmail()).isPresent()) {
            throw new EmailAlreadyExistsException("Email is already taken.");
        }
        Member newMember = Member.builder()
                .email(dto.getEmail())
                .password(bCryptPasswordEncoder.encode(dto.getPassword()))
                .build();
        return memberRepository.save(newMember).getId();
    }

    public class EmailAlreadyExistsException extends RuntimeException {
        public EmailAlreadyExistsException(String message) {
            super(message);
        }
    }
}
